Allow job_regex to match agains jobs to block
I have came across the job names not matching when using CircleCI matrix config.  For example you can do
runTests:
   matrix:
      parameters:
         test: [ 'a', 'b', 'c', 'd', 'e' ]

However,  When the CIRCLE_JOB name gets produced,  instead of runTests.  CircleCI changes the CIRCLE_JOB to "runTests-a"
This change should allow a regex to be optionally passed in so we can match against "^runTests*"