Use committer_date instead of workflow created at date
As summarized in this github issue: https://github.com/eddiewebb/circleci-queue/issues/95
Gather seems to have some issues with the way the original queue orb works
and calculates which workflows should be blocked and not.
This commit tries to tackle these issues.
More details around the problem and the solution proposed can be found in
the above issue (https://github.com/eddiewebb/circleci-queue/issues/95)