Update README to clarify API token and permissions required
Fixes #103.

Note that the CircleCI documentation is scant on information about what
permissions are required to query and cancel jobs, but in testing in the
UI at least, a user with Viewer access cannot cancel another user's
running job, suggesting that Contributor (write access) is required for
this orb to operate.